moved llm declaration
